refactor: extract stores to individual files
