+++ packages/app/src/components/sidebar/ObjectsPanel.tsx
@@ -1,4 +1,4 @@
-import { useDisplayEntityStore } from '@/store'
+import { useDisplayEntityStore } from '@/stores/displayEntityStore'
 import { cn } from '@/utils'
 import { FC } from 'react'
 import { IoCubeOutline } from 'react-icons/io5'
