+++ packages/app/src/stores/dialogStore.ts
@@ -1,0 +1,24 @@
+import { create } from 'zustand'
+import { immer } from 'zustand/middleware/immer'
+
+// ==========
+type DialogType =
+  | 'appInfo'
+  | 'blockDisplaySelect'
+  | 'itemDisplaySelect'
+  | 'exportToMinecraft'
+  | null
+type DialogState = {
+  openedDialog: DialogType
+  setOpenedDialog: (dialog: DialogType) => void
+}
+
+export const useDialogStore = create(
+  immer<DialogState>((set) => ({
+    openedDialog: null,
+    setOpenedDialog: (dialog) =>
+      set((state) => {
+        state.openedDialog = dialog
+      }),
+  })),
+)
