+++ packages/app/src/stores/modelDataStore.ts
@@ -1,0 +1,46 @@
+import { ModelDisplayPositionKey, ModelElement } from '@/types'
+import { create } from 'zustand'
+import { immer } from 'zustand/middleware/immer'
+
+// ==========
+type ModelData = {
+  textures: Record<string, string>
+  display: Record<
+    ModelDisplayPositionKey,
+    {
+      rotation?: [number, number, number]
+      translation?: [number, number, number]
+      scale?: [number, number, number]
+    }
+  >
+  elements: ModelElement[]
+}
+type ModelDataStoreState = {
+  modelData: Record<
+    string,
+    {
+      data: ModelData
+      isBlockShapedItemModel: boolean
+    }
+  >
+  setModelData: (
+    resourceLocation: string,
+    data: ModelData,
+    isBlockShapedItemModel: boolean,
+  ) => void
+}
+// 모델 데이터 캐시 저장소
+
+export const useModelDataStore = create(
+  immer<ModelDataStoreState>((set) => ({
+    modelData: {},
+    setModelData: (resourceLocation, data, isBlockShapedItemModel) =>
+      set((state) => {
+        const modelData = state.modelData[resourceLocation]
+
+        if (modelData == null) {
+          state.modelData[resourceLocation] = { data, isBlockShapedItemModel }
+        }
+      }),
+  })),
+)
